Move login functions from ClustersService to useClusterLogin
