+++ web/packages/teleterm/src/ui/services/clusters/clustersService.ts
@@ -25,20 +25,15 @@ import { Gateway } from 'gen-proto-ts/teleport/lib/teleterm/v1/gateway_pb';
 import {
   CreateAccessRequestRequest,
   CreateGatewayRequest,
-  PasswordlessPrompt,
   PromoteAccessRequestRequest,
   ReviewAccessRequestRequest,
 } from 'gen-proto-ts/teleport/lib/teleterm/v1/service_pb';
 import { useStore } from 'shared/libs/stores';
-import { isAbortError } from 'shared/utils/abortError';
+import { isAbortError } from 'shared/utils/error';
 import { pipe } from 'shared/utils/pipe';
 
 import { MainProcessClient } from 'teleterm/mainProcess/types';
-import {
-  CloneableAbortSignal,
-  cloneAbortSignal,
-  TshdClient,
-} from 'teleterm/services/tshd';
+import { cloneAbortSignal, TshdClient } from 'teleterm/services/tshd';
 import { getGatewayTargetUriKind } from 'teleterm/services/tshd/gateway';
 import { NotificationsService } from 'teleterm/ui/services/notifications';
 import { UsageService } from 'teleterm/ui/services/usage';
@@ -109,52 +104,6 @@ export class ClustersService extends ImmutableStore<types.ClustersServiceState> 
     });
   }
 
-  async loginLocal(
-    params: types.LoginLocalParams,
-    abortSignal: CloneableAbortSignal
-  ) {
-    await this.client.login(
-      {
-        clusterUri: params.clusterUri,
-        params: {
-          oneofKind: 'local',
-          local: {
-            user: params.username,
-            password: params.password,
-            token: params.token,
-          },
-        },
-      },
-      { abort: abortSignal }
-    );
-    // We explicitly use the `andCatchErrors` variant here. If loginLocal succeeds but syncing the
-    // cluster fails, we don't want to stop the user on the failed modal – we want to open the
-    // workspace and show an error state within the workspace.
-    await this.syncAndWatchRootClusterWithErrorHandling(params.clusterUri);
-    this.usageService.captureUserLogin(params.clusterUri, 'local');
-  }
-
-  async loginSso(
-    params: types.LoginSsoParams,
-    abortSignal: CloneableAbortSignal
-  ) {
-    await this.client.login(
-      {
-        clusterUri: params.clusterUri,
-        params: {
-          oneofKind: 'sso',
-          sso: {
-            providerType: params.providerType,
-            providerName: params.providerName,
-          },
-        },
-      },
-      { abort: abortSignal }
-    );
-    await this.syncAndWatchRootClusterWithErrorHandling(params.clusterUri);
-    this.usageService.captureUserLogin(params.clusterUri, params.providerType);
-  }
-
   async authenticateWebDevice(
     rootClusterUri: uri.RootClusterUri,
     {
@@ -178,99 +127,6 @@ export class ClustersService extends ImmutableStore<types.ClustersServiceState> 
         expectedDeviceIds: [],
       },
     });
-  }
-
-  async loginPasswordless(
-    params: types.LoginPasswordlessParams,
-    abortSignal: CloneableAbortSignal
-  ) {
-    await new Promise<void>((resolve, reject) => {
-      const stream = this.client.loginPasswordless({
-        abort: abortSignal,
-      });
-
-      let hasDeviceBeenTapped = false;
-
-      // Init the stream.
-      stream.requests.send({
-        request: {
-          oneofKind: 'init',
-          init: {
-            clusterUri: params.clusterUri,
-          },
-        },
-      });
-
-      stream.responses.onMessage(function (response) {
-        switch (response.prompt) {
-          case PasswordlessPrompt.PIN:
-            const pinResponse = (pin: string) => {
-              stream.requests.send({
-                request: {
-                  oneofKind: 'pin',
-                  pin: { pin },
-                },
-              });
-            };
-
-            params.onPromptCallback({
-              type: 'pin',
-              onUserResponse: pinResponse,
-            });
-            return;
-
-          case PasswordlessPrompt.CREDENTIAL:
-            const credResponse = (index: number) => {
-              stream.requests.send({
-                request: {
-                  oneofKind: 'credential',
-                  credential: { index: BigInt(index) },
-                },
-              });
-            };
-
-            params.onPromptCallback({
-              type: 'credential',
-              onUserResponse: credResponse,
-              data: { credentials: response.credentials || [] },
-            });
-            return;
-
-          case PasswordlessPrompt.TAP:
-            if (hasDeviceBeenTapped) {
-              params.onPromptCallback({ type: 'retap' });
-            } else {
-              hasDeviceBeenTapped = true;
-              params.onPromptCallback({ type: 'tap' });
-            }
-            return;
-
-          // Following cases should never happen but just in case?
-          case PasswordlessPrompt.UNSPECIFIED:
-            stream.requests.complete();
-            return reject(new Error('no passwordless prompt was specified'));
-
-          default:
-            stream.requests.complete();
-            return reject(
-              new Error(
-                `passwordless prompt '${response.prompt}' not supported`
-              )
-            );
-        }
-      });
-
-      stream.responses.onComplete(function () {
-        resolve();
-      });
-
-      stream.responses.onError(function (err: Error) {
-        reject(err);
-      });
-    });
-
-    await this.syncAndWatchRootClusterWithErrorHandling(params.clusterUri);
-    this.usageService.captureUserLogin(params.clusterUri, 'passwordless');
   }
 
   /**
